% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalmle.R
\name{equalmle}
\alias{equalmle}
\title{This provides MLEs for the parameters of a one-way ANCOVA model when the treatment effects are identical.}
\usage{
equalmle(data1, data2, k, q)
}
\arguments{
\item{data1}{a real data set}

\item{data2}{a real data set}

\item{k}{a positive integer}

\item{q}{a positive integer}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=4;q=3
N=c(20,10,10,20,20,10,10,20,20,10,10,20);S=c(1,1,2,1,1,2,3,1,2,4,6,3,2,4)
g=NULL
for(i in 1:(k*q))
{
 g[[i]]=rnorm(N[i],0,sqrt(S[i]))
}
X=g
G2=NULL
N=c(20,10,10,20);a=c(1,2,3,4)
for(i in 1:k)
{
 G2[[i]]=rnorm(N[i],a[i],sqrt(S[i]))
}
Y=G2
equalmle(Y,X,k,q)
}
