% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneWayANCOVAMLE.R
\docType{package}
\name{OneWayANCOVAMLE}
\alias{OneWayANCOVAMLE}
\title{Find MLEs of parameters of a one-way ANCOVA model.}
\description{
This provides MLEs for the parameters of a one-way ANCOVA model in the cases where the treatment effects are monotonically increasing and identical under heterogeneous error variances.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
